requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.13.0
APScheduler==3.10.4
asyncpg==0.31.0
attrs==26.1.0
bitarray==3.8.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.2
cryptography==46.0.7
curl_cffi==0.15.0
cytoolz==1.1.0
discord-protos==0.0.2
discord.py==2.7.1
discord.py-self==2.1.0
distro==1.9.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==6.0.0b1
fastapi==0.115.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
metar==2.0.1
multidict==6.7.1
numpy==1.26.4
packaging==26.0
parsimonious==0.10.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
propcache==0.4.1
protobuf==7.34.1
psycopg2-binary==2.9.9
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.6
py_order_utils==0.3.2
pybit==5.14.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.20.0
pytest==9.0.2
python-dotenv==1.0.1
python-metar==2.0.1
python-telegram-bot==21.5
pytz==2026.1.post1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
rlp==4.1.0
scipy==1.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
toolz==1.1.0
types-requests==2.33.0.20260408
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
web3==7.15.0
websocket-client==1.9.0
websockets==13.0
yarl==1.23.0
